requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
fastapi==0.115.0
uvicorn==0.30.1
joblib==1.4.2
matplotlib==3.8.4
mlflow==2.10.0
pyyaml==6.0.1
pytest==8.3.1